Add Unit.getRadar skeleton
Add the skeleton code for the dcs `Unit.getRadar` method.

In doing so it became quickly apparent that one .proto file would get
unwieldy quickly so split it out into smaller files. I have created a
common file to hold the enums and objects I expect to be returned from
all manner of methods.

Although I split the .proto file I kept all the methods in the base
`dcs` protobuf package. Although there was a temptation to create
sub-packages I do not believe it would work well when dealing with
methods on objects that inherit since I do not believe the client cares
about which object the method is actually being called on.

This might change as I try to implement some methods from the object
hierarchy so it is best to experiment with this now since this changes
the generated code in both the server and the clients so we want to do
this before things get too painful to change.