Use full Weapon information
Replace the simple int for weapon name with more useful information such
as what type of weapon it is.

The `getName` function on a weapon returns an integer instead of a
string. To me this *looks* like it should be an ID rather than a name
but will leave as is because second guessing ED will probably lead to
madness.