Add World.getMarkPanels API
Add the World.getMarkPanels API so that we can implemented all the
MarkPanel APIs.

I have only tested in single player where it appears that the initiator
is _always_ the player even if the MarkPoint was created via scripting.

It also looks like, in SP at least, that "MarkToAll" API will result in
the markpoint being the "Blue" coalition.

I haven't yet had the opportunity to test on an MP server.