Add Speed and Heading to unit information
Add the speed and heading of a unit to the default information exported.
These two values are used a lot in map visualisations and in OverlordBot
so we are exporting them by default rather than making a separate API.

We can add these as separate APIs also if it turns out to be needed.