Make some triggers relative to ground.
Make the triggers for IlluminationBomb, Smoke & Flares relative to
ground so they are more useful.

Smoke & Flares will always be at ground level and ignore the altitude
parameter.

Illumination bombs will use the altitude parameter as an AGL value so
that the client doesn't need to calculate ground level and having an
Illumination bomb at 13,000ft is not very useful.