Follow linting advice
Eat crow and fix the things I thought I was being smart about and
instead follow the recommendations of soulless linting programs.

We are not following the suggestions for versioning the APIs at the end
yet since are are still pre-stable.

We will fix the enum warnings regarding the usage of `0` in a future
commit

Files formatted using clang-format