Proof of Concept for Enum renumbering
gRPC does not recommend zero indexed Enums but DCS uses them a lot.
Therefore this commit is a proof-of-concept for the easiest workaround I
can think of.

Pros: Simple.
Cons: all lua using these enums will have to remember to -1 when the DCS
enum is zero indexed (which is not all of them).