Add initial `AddGroup` implementation
So far this implementation will create a ground group with a number of
ground units and return null in the gRPC response for some reason.

Many options are still hardcoded but what is currently implemented is
enough to spawn a static SAM site such as a NASAM site and have it
act as expected.