Add logging around config file writing
Add logging regarding the writing of the `Data\dcs-grpc.lua`
file and change the check to make sure that `f` is not nill rather
than there is a definite error as this appears to be the safer, more
idiomatic, way of doing it.

This should help a bit with debugging issues some users appear to
be having.