chat messages are synchronized to event stream
Observers of the simulation events and the chat messages (and other
server side events) would be difficult to synchronize the chronological
order unless the process creates a Rx stream which would require latency
to be introduced to synchronize the events.

Taking inspiration from PR#88, emitting net.* based events on a single
stream would ensure we have a consistent model with zero latency.