use config file as only source for settings
Before this commit it was possible to set settings either:
1) before the `dofile` inside of `MissionScription.lua`,
2) before `GRPC.load()` inside of a mission, and
3) inside of the config file at `Saved Games\DCS\Config\dcs-grpc.lua`.

This commit
a) removes (1),
b) documents the limitations of (2),
c) splits `Hooks/DCS-gRPC.lua` into two parts, and
c) removes all unnecessary `_G.` prefixes on globals (in Lua code; just
   a cleanup as a side-effect of the refactoring).

Reducing the amount of locations that settings can be configured (a)
simplifies the code and reduces the amount things that need to be
tested. More importantly, it removes the need of the temporary
`Data/dcs-grpc.lua` file (which is removed).

With (c) it is now easier to develop the hooks part of DCS-gRPC as it
allows a setup where the actual hook logic (now in
`lua\DCS-gRPC\grpc-hook.lua`) can be loaded from a local repository.
`lua\Hooks\DCS-gRPC.lua` is still there, but it should rarely change
during development as it is only responsible for loading the config and
then executing `lua\DCS-gRPC\grpc-hook.lua` (where the actual hook logic
is implemented).