Add MissionCommand API support
Add support for the `AddMissionCommand`, `AddMissionCommandSubMenu` and
`RemoveMissionCommandItem` APIs along with their Coalition and Group
variants which allow us to create F10 radio command menu entries and
have them emit events to DCS-gRPC clients.

Additionally it adds new CommandEvents which are triggered for
any execution of a command via the F10 radio menu by the players.