Update README.md setup instructions (#99)
* Update README.md setup instructions

Update the README.md instructions for setting up DCS-gRPC based on
experience dealing with questions from people installing it and
confused where and what they have to do.

This involves duplicating information which kind of sucks from a 
maintenance point of view but clarifies things. I have also removed
information on the `autostart` since this also caused confusion with
the explanation. All the installers need to know is if they set it
to true or not.

I am also not talking about the possibility of having the config file
with autostart = false and still doing the `GRPC.load()` in the miz
since it again causes confusion

Also add hard newlines at around the 120 column point for nicer
formatting.