SendChat and SendChatTo
Initial implementation of the `net.*` namespace with the:
- `send_chat_to`
- `send_chat`

`send_chat_to` does not implement the DCS API parameter of `fromPlayerId` due
to the social impact of a player being emulated is more severe than a breached system
and hence why the parameter has been redacted. If there is a valid use case, please
do feel free to present your use case and why alternatives do not meet the
user requirements. See PR #94

Added `net` as global read for lua linter