GetPlayers returns all connected player attributes
Getting the player info based on the players connected to the server. This is a
combined loop of `net.get_player_list` and `net.get_player_info`.

Notes about the player information which differs from official documentation.
- locale is an undocumented attribute / would serve well for multilingual cases
- ipAddr is actually `{address}:{port}`

It was decided to perform the compound variation of the API. In theory, it should
work effectively but needs to be verified by densely populated servers.