Add C# headers
Add headers to the proto files that specify the namespace the generated
C# files will live under. This is part of work to create a simple C#
nuget package so clients to not have to re-invent the wheel in their own
appications.

There are already `Dcs.*` nuget packages so to avoid confusion and
conflict I am putting these under `RurouniJones` as a unique namespace
since we don't have a company or any name that is particularly unique
and I will be publishing other C# DCS related packages so it is as good
a place as any.