Add more detail to `GroupCommand` event.
Replace `groupName` field in the `GroupCommand` event with all the group
details as exposed by the group exporter (currently id, name, coalition
and category).

This change was made based on experience writing a client that processes
these events where only having the groupName was a limitation.

This change breaks backwards compatibility with version 0.4.0 where the
`GroupCommand` event was first added.