Add Static Objects API (`AddStaticObject` and `AddLinkedStatic`)
Implementation of `addStaticObject` via the gRPC interface.

Two API's were added as the behavior is slightly different that it did
not need to justify an overloaded/oneof complex body.

The cargo type is free text so to avoid the requirements of DCS-gRPC
installations and clients to be coupled to any DCS updates / new units /
 renames

* Static positioning
* Linked unit
* Cargo handling