Add Mission Management APIs
Add mission management APIs that allow clients to load and restart
missions. The `LoadNextMission` API is apparently supposed to return
false when the last mission in the mission list has been reached but DCS
is returning true all the time.

The APi to reload the current mission is not natively provided
by DCS but is simple to implement and will be useful for admins I think.