Make Hit/Kill event weaponName field optional
Not all hit and kill events include a `weaponName` depending on what
happened (For example one aircraft hitting another without the use of
an intermediate weapon). This was causing many deserialization errors in
the production logs of servers. Therefore we will make this field
optional.

I think that longterm we can come up with a better solution by removing
the `weaponName` field entirely and only ever populating the `Weapon`
field instead in lua. However this is a short-term fix for the next
release.