Return the full parent group with the unit
When returning a unit object we will return the full group as its own
object instead of embedding some group attributes in the unit object.
This makes it easier for clients to parse streaming events without
having to make separate calls to get group information or require them
to keep a local copy in state that then needs to be managed.