Add API Key Authentication (#269)
Added API Key authentication. 

Due to a recurring question about authentication of clients I've implemented a Interceptor layer to the tonic server to check all calls for valid api keys. 

example config: 
```
auth.enabled = false --defaults to false
auth.tokens = {
    { client = "test", token = "Sometest" },
    { client = "another client", token = "Some other test" }
}
```
`auth.tokens` is a table of auth keys with their client name. 
There can be a "default" client or a single key for all clients, but this is up to configuration. 
There can be as many client keys as needed. 

In the debug log the client that authenticates is logged. 

### Performance
Performance wise there is no notable difference. 

### Possible future features

* Possibly in the future "expiration_date" can be added to automatically revoke issues, but I didn't think that was needed for a first implementation. 

* Add per client `eval` authorisation